Disable 'continue' button when there are no saves
