+++ Assets/Scripts/Components/EntryHover.cs
@@ -5,11 +5,13 @@ using UnityEngine;
 namespace Menu {
     public class EntryHover : MonoBehaviour {
         public Color HoverColor = Color.yellow;
+        public Color DisabledColor = new Color(0.49f, 0.49f, 0.49f);
 
         private AudioSource _audioHoverSource;
         private Renderer _renderer = new Renderer();
         private Color _inactiveColor = Color.gray;
         private bool _isHovering;
+        protected bool IsEnabled = true;
 
         protected virtual void OnClick() {}
 
@@ -20,16 +22,20 @@ namespace Menu {
             this.OnClick();
         }
 
-        private void Start() {
+        protected void Start() {
             this._audioHoverSource = this.GetComponent<AudioSource>();
             this._renderer = this.GetComponent<Renderer>();
             Cursor.visible = true;
+
+            if (!this.IsEnabled) {
+                this._renderer.material.color = this.DisabledColor;
+            }
         }
 
         private void OnMouseEnter() {
             // we need to ensure this is not being
             // called multiple times during an update
-            if (this._isHovering) {
+            if (this._isHovering || !this.IsEnabled) {
                 return;
             }
 
@@ -46,9 +52,6 @@ namespace Menu {
 
             this._renderer.material.color = this._inactiveColor;
             this._isHovering = false;
-        }
-
-        private void Update() {
         }
     }
 }
