+++ Assets/Scripts/Components/EntryHover.cs.meta
