+++ Assets/Scripts/Managers.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 30ce5b3f292b48b18ae8b3659910937d
+timeCreated: 1539439855
