+++ Assets/Scripts/Managers/BackgroundGameManager.cs
@@ -1,0 +1,26 @@
+﻿using UnityEditor;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace Managers {
+    public class BackgroundGameManager : MonoBehaviour {
+        public SceneAsset DebugScene;
+        private static bool _created;
+
+        private void Awake() {
+            if (!_created) {
+                DontDestroyOnLoad(this.gameObject);
+                _created = true;
+            }
+            else {
+                Destroy(this.gameObject);
+            }
+        }
+
+        private void FixedUpdate() {
+            if (Input.GetKey (KeyCode.F12)) {
+                SceneManager.LoadScene(this.DebugScene.name);
+            }
+        }
+    }
+}
