+++ Assets/Scripts/Managers/SaveGameManager.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 57d739d928b3455484aeb2607439212b
+timeCreated: 1539439892
